Fix issue with changing STIX IDs
This commit fixes an issue where the unique IDs in the generated STIX
were not consistent, meaning the same object would get a different
random ID each time. This made it challenging for downstream tools to
work with the published STIX bundle.

The implementation uses UUID5 to generate consistent UUIDs based on a
domain that was selected for this purpose and the ATLAS IDs from the
YAML data (e.g. AML.T0000). These IDs are consistent and do not change
for the life of the object. For relationships, a string is constructed
that combines IDs from both objects in the relationship (e.g.
AML.T0000.000-subtechnique-of-AML.T0000). In places where there wasn't
an equivalent object in the ATLAS YAML, a consistent string was chosen.

This commit also uses the created and modified dates from the ATLAS YAML
file to set the timestamps in STIX. Previously, the default of the
current time was being used for both. The exception is the timstamp for
the ATLAS matrix because that is not tracked in the ATLAS YAML.